generate signatures for reloadable functions automatically via a proc-macro

+++ src/lib.rs
@@ -2,6 +2,14 @@
 
 /*!
 
+[![Crates.io](https://img.shields.io/crates/v/hot-lib-reloader)](https://crates.io/crates/hot-lib-reloader)
+[![](https://docs.rs/structopt/badge.svg)](https://docs.rs/hot-lib-reloader)
+[![License](https://img.shields.io/crates/l/hot-lib-reloader?color=informational&logo=mit)](/LICENSE.md)
+
+<!--
+[![Crates.io](https://img.shields.io/crates/d/hot-lib-reloader)](https://crates.io/crates/hot-lib-reloader)
+ -->
+
 A simple crate around [libloading](https://crates.io/crates/libloading) that can be used to watch Rust libraries (dylibs) and will reload them again when they have changed.
 Useful for changing code and seeing the effects without having to restart the app.
 
@@ -35,7 +43,7 @@ And `lib/lib.rs`
 
 ```
 #[no_mangle]
-pub extern "C" fn do_stuff() {
+pub fn do_stuff() {
     println!("doing stuff");
 }
 ```
@@ -52,17 +60,25 @@ edition = "2021"
 
 [dependencies]
 lib = { path = "../lib" }
-hot-lib-reloader = "0.3.0"
+hot-lib-reloader = "0.4.0"
 ```
 
 You can then define and use the lib reloader like so:
 
 ```no_run
-hot_lib_reloader::define_lib_reloader!(
-    MyLibLoader("target/debug", "lib") {
-        fn do_stuff() -> ();
+hot_lib_reloader::define_lib_reloader! {
+    unsafe MyLibLoader {
+        // Will look for "liblib.so" (Linux), "lib.dll" (Windows), ...
+        lib_name: "lib",
+        // Where to load the reloadable functions from,
+        // relative to current file:
+        source_files: ["../../lib/src/lib.rs"]
+        // You can optionally specify manually:
+        // functions: {
+        //     fn do_stuff();
+        // }
     }
-);
+}
 
 fn main() {
     let mut lib = MyLibLoader::new().expect("init lib loader");
@@ -115,9 +131,16 @@ cargo watch -w bin -x run
 
 A change that you now make to `lib/lib.rs` will have an immediate effect on the app.
 
-*/
 
-mod macros;
+# More examples
+
+Examples can be found at [rksm/hot-lib-reloader-rs/examples](https://github.com/rksm/hot-lib-reloader-rs/tree/master/examples).
+
+- [minimal](https://github.com/rksm/hot-lib-reloader-rs/tree/master/examples/minimal): Bare-bones setup.
+- [reload-feature](https://github.com/rksm/hot-lib-reloader-rs/tree/master/examples/reload-feature): Use a feature to switch between dynamic and static version.
+- [bevy](https://github.com/rksm/hot-lib-reloader-rs/tree/master/examples/bevy): Shows how to hot-reload bevy systems.
+
+*/
 
 use libloading::Library;
 use libloading::Symbol;
@@ -134,6 +157,8 @@ use std::thread;
 use std::time::Duration;
 use std::time::Instant;
 
+pub use hot_lib_reloader_macro::define_lib_reloader;
+
 pub struct LibReloader {
     changed: Arc<atomic::AtomicBool>,
     changed_at: Arc<Mutex<Option<Instant>>>,
@@ -147,6 +172,12 @@ impl LibReloader {
         lib_dir: impl AsRef<Path>,
         lib_name: impl AsRef<str>,
     ) -> Result<Self, Box<dyn Error>> {
+        // find the target dir in which the build is happening and where we should find
+        // the library
+        let lib_dir = find_file_or_dir_in_parent_directories(lib_dir.as_ref())?;
+        log::debug!("found lib dir at {lib_dir:?}");
+
+        // sort out os dependent file name
         #[cfg(target_os = "macos")]
         let (prefix, ext) = ("lib", "dylib");
         #[cfg(target_os = "linux")]
@@ -154,15 +185,21 @@ impl LibReloader {
         #[cfg(target_os = "windows")]
         let (prefix, ext) = ("", "dll");
         let lib_name = format!("{prefix}{}", lib_name.as_ref());
-        let watched_lib_file = lib_dir.as_ref().join(&lib_name).with_extension(ext);
+
+        let watched_lib_file = lib_dir.join(&lib_name).with_extension(ext);
         let loaded_lib_file = lib_dir
-            .as_ref()
             .join(format!("{lib_name}-hot"))
             .with_extension("dll");
-        // We don't load the actual lib because this can get problems e.g. on Windows
-        // where a file lock would be held, preventing the lib from changing later.
-        std::fs::copy(&watched_lib_file, &loaded_lib_file)?;
-        let lib = Some(unsafe { Library::new(&loaded_lib_file) }?);
+
+        let lib = if watched_lib_file.exists() {
+            // We don't load the actual lib because this can get problems e.g. on Windows
+            // where a file lock would be held, preventing the lib from changing later.
+            std::fs::copy(&watched_lib_file, &loaded_lib_file)?;
+            Some(unsafe { Library::new(&loaded_lib_file) }?)
+        } else {
+            log::debug!("library {watched_lib_file:?} does not yet exist");
+            None
+        };
 
         let lib_loader = Self {
             watched_lib_file: watched_lib_file.clone(),
@@ -205,8 +242,13 @@ impl LibReloader {
         if let Some(lib) = lib.take() {
             lib.close()?;
         }
-        std::fs::copy(watched_lib_file, loaded_lib_file)?;
-        self.lib = Some(unsafe { Library::new(&self.loaded_lib_file) }?);
+
+        if watched_lib_file.exists() {
+            std::fs::copy(watched_lib_file, loaded_lib_file)?;
+            self.lib = Some(unsafe { Library::new(&self.loaded_lib_file) }?);
+        } else {
+            log::warn!("trying to reload library but it does not exist");
+        }
 
         Ok(())
     }
@@ -282,8 +324,8 @@ impl LibReloader {
                             thread::sleep(debounce);
                         }
                     }
-                    Ok(_change) => {
-                        // dbg!(change);
+                    Ok(change) => {
+                        log::trace!("file change event: {change:?}");
                     }
                     Err(err) => {
                         log::error!(
@@ -318,5 +360,36 @@ impl Drop for LibReloader {
         if self.loaded_lib_file.exists() {
             let _ = std::fs::remove_file(&self.loaded_lib_file);
         }
+    }
+}
+
+/// Try to find that might be a relative path such as `target/debug/` by walking
+/// up the directories, starting from cwd. This helps finding the lib when the
+/// app was started from a directory that is not the project/workspace root.
+fn find_file_or_dir_in_parent_directories(
+    file: impl AsRef<Path>,
+) -> Result<PathBuf, Box<dyn Error>> {
+    let mut file = file.as_ref().to_path_buf();
+    if !file.exists() && file.is_relative() {
+        if let Ok(cwd) = std::env::current_dir() {
+            let mut parent_dir = Some(cwd.as_path());
+            while let Some(dir) = parent_dir {
+                if dir.join(&file).exists() {
+                    file = dir.join(&file);
+                    break;
+                }
+                parent_dir = dir.parent();
+            }
+        }
+    }
+
+    if file.exists() {
+        Ok(file)
+    } else {
+        Err(std::io::Error::new(
+            std::io::ErrorKind::NotFound,
+            format!("file {file:?} does not exist"),
+        )
+        .into())
     }
 }
